requirements.txt
flask==2.2.5
scikit-learn
nltk
pdfplumber; platform_system != 'Windows'
python-docx
sentence-transformers
